Make return types of `_parent()` and `_root()` nullable
`_parent` and `_root` are `null` when you invoke an external type (i.e.
a type that is not defined inside the .ksy spec from which you call it).
This was implemented in
<https://github.com/kaitai-io/kaitai_struct_compiler/pull/283>. However,
the return type `Struct` doesn't allow `null`, only its nullable variant
`?Struct` does.

Without this change, the NestedTypesImport test fails with:

```
$ ./docker-ci -t php -i 8.3
...
2) Kaitai\Struct\Tests\NestedTypesImportTest::testNestedTypesImport
TypeError: Kaitai\Struct\Struct::_parent(): Return value must be of type Kaitai\Struct\Struct, null returned

/runtime/php/lib/Kaitai/Struct/Struct.php:31
/tests/spec/php/NestedTypesImportTest.php:15
```

Since the nullable return types were added in PHP 7.1 (see
https://www.php.net/manual/en/migration71.new-features.php#migration71.new-features.nullable-types),
this increases the minimum supported PHP version from 7.0 to 7.1, but I
don't think this is a big deal since PHP 7.0 is EOL for a long time
(since 2019-01-10, see https://endoflife.date/php). If we wanted to keep
supporting PHP 7.0, I think we would have to remove the return type
hint, but I don't see a reason to do that (at this point, the oldest PHP
version "supported for critical security issues only" is PHP 8.1, so we
could probably even drop support for anything older than that).